Convert to using free arrows. (#6)
* Add free arrow definition.

* Convert to using free arrows.

As a consequence to this, we cannot perform the introspection on
the `Flow` type which was used to build e.g. the pretty printer. To
deal with this, we introduce the `Diagram` type, which is a 'static'
representation to which we can compile a `Flow` and use for presentation
purposes.

A couple of things have been "lost" as part of this change:

- `Name` is now `Named` and may only name pure computations, not
  arbitrary flows.
- `Catch` has gone. Like `Named`, this involves non-arrow based
  composition at the term level. I think if we were to bring it back,
  we should do so by adding it to the core arrow structure, as in e.g.
  https://hackage.haskell.org/package/arrows-0.4.4.1/docs/Control-Arrow-Operations.html

* Reintroduce working exceptions.

We make a couple of changes:
- Use a generic exception type rather than 'String'!
- Use Control.Monad.Catch from 'exceptions', since most people seem
	to recommend this.

Formatting.

WIP Redis update.