Add content addressable store
After an item is completed its hash is calculated and the contents
are moved to a path derived from that hash. If that hash already
exists in the store, then the new version is discarded.
A symbolic link will point from the input's content hash to the result.

In future a GC mechanism and back-references from items to the
corresponding input hashes should be added.