Revert "Simplify interaction between store and executor"
This reverts commit d15159e26a5238419857ab3d58033414b2ca84d9.

The logic where the runner was responsible for creating all 'pending'
items could also lead to problems, where the runner would see a pending
item but no task was in the coordinator. This commit adds more exhaustive
testing for the status of an item; the executor is responsible for creating
pending items, but the runner will verify with the coordinator that they truly
are pending. If not, it will clean them up and resubmit them.

There is still a case where things could go wrong; if the executor dies, but
the task remains as 'running' in the coordinator. There's no easy solution to
this, save for introducing timeouts to tasks.