Cancel all async actions if one of them dies.
If running things in parallel, we can have a situation where one dies and throws
an exception up to the main runner, whilst leaving other threads running in the
background. This is undesirable, because now the whole program may die without
cleaning up those child threads. This change alters the behaviour of 'AsyncA'
such that if one action dies, all of them will (and hence hopefully run clean
up actions).

If the desire is to allow some actions to fail, then suitable exception handling
logic must be written inside each branch, such as to not cause the exception to
be handled by the Async.