Produce full output in executor
Previously, the file watcher was cancelled as soon as the inner task
completed, potentially not giving it the chance to print out the last
few lines of output. Now, an MVar is used to communicate to the file
watch that the inner task completed. The file watcher can then still
read the remainder of the output and terminate on its own terms.