Move funflow package to top-level.
The overarching reason is this: the funflow dir really does need
a README at its root. Previously it did not, because the README was at
repo top-level. But this causes problems. In particular, it means the
funflow dir is no longer self-contained. More importantly, you can't
include the README in package sdists uploaded to Hackage.

With funflow at top-level, we have top-level README, which is also the
funflow readme. This is natural because the top-level README is about
the top-level package, aka funflow.