Additional metadata in external tasks (#77)
* Pass original input to metadata writer

The `ExternalTask` record alone does not contain much information that
would be useful to generate additional user-defined metadata.

* export ExternalProperties and MDWriter

* Store ByteString in additional metadata

For more flexible metadata storage. E.g. encoded Haskell data-types.
The SQLite database that stores the metadata is capable of storing
arbitrary binary blobs.

* Allow to write metadata on non-pending items

Previously, metadata could only be modified while a task was pending.
Since the additional metadata storage happens in the flow executor, and
not the external task executor, this cannot be guaranteed.