Fix: Build failures in GHC 8.6 (including nix build) (#135)
* Fix: Build failures in GHC 8.6 (including nix build)

This was due to a partial pattern match. Since the use of this pattern match is
meant to represent an unfailable case, we simply use a wildcard bind here
instead. If this fails in some case, it's not the end of the world, since the
task shouldn't be in the store and an error will be thrown there.

* Update funflow/src/Control/Funflow/Exec/Simple.hs

Co-Authored-By: nc6 <nick@topos.org.uk>