Re-organize nix config (#43)
Re-organized nix config to use overlays to grab our various
packages produced by haskell.nix and export the relevant ones as
top-level attributes in a new default.nix that lives at the repository root. 
The default.nix in the nix directory now exports the nixpkgs set that
the project uses. Also updated the method for how the API docs get built
to use relative paths in the haddock docs for packages in this repo 
instead of absolute paths.