Add client library for docker (#52)
Added a new library to the repository, docker-client, which provides methods for running and extracting data from docker containers via the Docker Engine API as opposed to the docker command line client. This provides us with tighter control on the commands we submit to docker and allows for easier processing of command outputs.  This library allows us to address the permissions issues noted in #45.