Add Tutorial using Tensorflow + Docker Tasks (#55)
Added a tutorial notebook which uses DockerTasks to train and run a Tensorflow image classification model. Also adjusted the nix build for tutorials to create an executable script which runs the tutorial notebooks instead of running them within the nix build. This allows them to (impurely) access services on the host like docker, which run over a unix socket which isn't available in the nix build's sandbox.