Update docker-client to use "latest" image by default. (#63)
Modified the behavior of defaultContainerSpec to add the "latest" tag to images if a tag or digest is not specified. Users who are interested in pulling all images in a repo (the way the Docker Engine API does by default) can still do so by manually constructing a ContainerSpec without a tag/digest. Also added a couple of related helper functions for modifying Docker image strings.